changed: split gl and dx implementation
while some amount of code is shared, it
does not justify the spaghetti
